backend/graph-services/extractor-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
langchain==0.1.0
langchain-openai==0.0.2
pydantic-ai==0.0.8
python-dotenv==1.0.0
python-multipart==0.0.6
